requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0
